Abstracts timer to seperate file
